[WFLY-6143] Update ejb3 model to 6.0
