+++ ejb3/src/main/java/org/jboss/as/ejb3/subsystem/EJB3Subsystem50Parser.java
@@ -33,10 +33,7 @@ import static org.jboss.as.controller.parsing.ParseUtils.requireNoNamespaceAttri
 import static org.jboss.as.controller.parsing.ParseUtils.unexpectedAttribute;
 import static org.jboss.as.controller.parsing.ParseUtils.unexpectedElement;
 import static org.jboss.as.ejb3.subsystem.EJB3SubsystemModel.APPLICATION_SECURITY_DOMAIN;
-import static org.jboss.as.ejb3.subsystem.EJB3SubsystemModel.CLASS;
 import static org.jboss.as.ejb3.subsystem.EJB3SubsystemModel.IDENTITY;
-import static org.jboss.as.ejb3.subsystem.EJB3SubsystemModel.MODULE;
-import static org.jboss.as.ejb3.subsystem.EJB3SubsystemModel.SERVER_INTERCEPTORS;
 import static org.jboss.as.ejb3.subsystem.EJB3SubsystemModel.SERVICE;
 
 import java.util.Collections;
@@ -85,10 +82,6 @@ public class EJB3Subsystem50Parser extends EJB3Subsystem40Parser {
             }
             case ENABLE_GRACEFUL_TXN_SHUTDOWN: {
                 parseEnableGracefulTxnShutdown(reader, ejb3SubsystemAddOperation);
-                break;
-            }
-            case SERVER_INTERCEPTORS: {
-                parseServerInterceptors(reader, ejb3SubsystemAddOperation);
                 break;
             }
             default: {
@@ -337,50 +330,5 @@ public class EJB3Subsystem50Parser extends EJB3Subsystem40Parser {
             }
         }
         return staticDiscovery;
-    }
-
-    protected void parseServerInterceptors(final XMLExtendedStreamReader reader, final ModelNode ejbSubsystemAddOperation) throws XMLStreamException {
-        final ModelNode interceptors = new ModelNode();
-
-        requireNoAttributes(reader);
-        while (reader.hasNext() && reader.nextTag() != XMLStreamConstants.END_ELEMENT) {
-            switch (EJB3SubsystemXMLElement.forName(reader.getLocalName())) {
-                case INTERCEPTOR: {
-                    parseInterceptor(reader, interceptors);
-                    break;
-                }
-                default: {
-                    throw unexpectedElement(reader);
-                }
-            }
-        }
-        requireNoContent(reader);
-
-        ejbSubsystemAddOperation.get(SERVER_INTERCEPTORS).set(interceptors);
-    }
-
-
-    protected void parseInterceptor(final XMLExtendedStreamReader reader, final ModelNode interceptors) throws XMLStreamException {
-        final ModelNode interceptor = new ModelNode();
-        for (int i = 0; i < reader.getAttributeCount(); i++) {
-            requireNoNamespaceAttribute(reader, i);
-            final String value = reader.getAttributeValue(i);
-            switch (EJB3SubsystemXMLAttribute.forName(reader.getAttributeLocalName(i))) {
-                case MODULE: {
-                    interceptor.get(MODULE).set(value);
-                    break;
-                }
-                case CLASS: {
-                    interceptor.get(CLASS).set(value);
-                    break;
-                }
-                default: {
-                    throw unexpectedAttribute(reader, i);
-                }
-            }
-        }
-        requireNoContent(reader);
-
-        interceptors.add(interceptor);
     }
 }
