+++ ejb3/src/main/java/org/jboss/as/ejb3/subsystem/EJB3SubsystemNamespace.java
@@ -43,7 +43,8 @@ public enum EJB3SubsystemNamespace {
     EJB3_2_0("urn:jboss:domain:ejb3:2.0"),
     EJB3_3_0("urn:jboss:domain:ejb3:3.0"),
     EJB3_4_0("urn:jboss:domain:ejb3:4.0"),
-    EJB3_5_0("urn:jboss:domain:ejb3:5.0");
+    EJB3_5_0("urn:jboss:domain:ejb3:5.0"),
+    EJB3_6_0("urn:jboss:domain:ejb3:6.0");
 
 
     private final String name;
